Create search module
